+++ src/Tracking/Requests/TrackingRequest.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace SmartDato\GlsItaly\Tracking\Requests;
+
+use Saloon\Enums\Method;
+use Saloon\Http\Request;
+
+abstract class TrackingRequest extends Request
+{
+    protected Method $method = Method::GET;
+
+    public function resolveEndpoint(): string
+    {
+        return '';
+    }
+}